backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==11.0.0
pdf2image==1.16.3
pypdf==4.1.0
opencv-python==4.10.0.84